Use consistent strategy for handling exceptions in auth controller
There was a bug in generateJwt which was causing it to raise an
exception, which was not being caught. Unfortunately, it's not possible
to encode in the type system that a function can throw, so it seems like
a good idea to code defensively and assume _any_ function can throw. To
that end, this commit adopts the strategy of putting everything in the
happy path within the `try` block and then within the `catch` we inspect
the error to determine what type it is and act accordingly.