Add JWT support for auth on server (#79)
* Pin to exact versions of dependencies

* Move @types/* packages into devDependencies

* Add auth packages

* Remove PORT env var which is specified in docker-compose.yml

* Return JWT on log in

Also includes some status code and error cleanup.

* Retrieve user id from logged in user's JWT

* Clean up

* Move files into appropriate folders
* Share relevant types
* Add more precise types (remove `any`s)
* Lint

* Fix generating JWT

The `sub` property seems to be required

* Fix exception handling for sending verification token

The body of the try can also throw exceptions not when an unhandled
error occurs but when a 400 level status code should be returned. We
need to check the type of exception in the catch to make sure we do the
right thing.

* Use consistent strategy for handling exceptions in auth controller

There was a bug in generateJwt which was causing it to raise an
exception, which was not being caught. Unfortunately, it's not possible
to encode in the type system that a function can throw, so it seems like
a good idea to code defensively and assume _any_ function can throw. To
that end, this commit adopts the strategy of putting everything in the
happy path within the `try` block and then within the `catch` we inspect
the error to determine what type it is and act accordingly.

* Revert "Add auth packages"

This reverts commit a2051003173d583ee7e6ed6d0225f51a3b609c13.

* Remove unused passport-local dependency

The official docs suggest using `passport-local` but this password
strategy is so dead-simple that we ended up not using a library for it.

* Update server/yarn.lock

* Require authentication for all endpoints in user controller

* Lint

* Reformat respecting prettier config