Use chamber as a template when setting number of districts
Always referencing a chamber to find the number of districts complicates
things by having to worry about versions of chambers as the number of
districts change over time. This switches approaches to instead use the
chamber as a template, retaining the number of districts on the project
itself which is copied over from the chamber configuration or set as
custom.