Add .eslintrc.js for 'client'
Adds an eslint config that attempts to more or less replicate the rules
we were enforcing with tslint.

This was *not* made using `tslint-to-eslint-config`, because that produced
an incredibly long config file that also needed a lot of help to even run.
It seems to be too concerned with exactly replicating the tslint rules on
one hand (even though many of those would have been default choices that
would be better imported by accepting the equivalent eslint defaults), and
also not as automatic as advertised.