Fix performance of GeoUnitHierarchy group function
This was scaling in overhead in a greater than linear fashion, making it
take an unacceptably long time for larger states.

By precollecting the data so each geounit is mapped by its parent key,
the lookups to determine the children for a given geounit are only done,
instead checking for every geounit.

Closes #152