Switch 'manage' from tslint to eslint
Removes tslint and installs eslint for the 'manage' package.
The .eslintrc.js file is based on the one for 'server', modified first
to reflect the differences between the old 'tslint.json' files (as much
as I could) then to turn off a couple rules that clearly weren't there
in tslint because there were lots of existing violations of the eslint
rule.