Work around issue with eslint exceptions for 'functional' rules
Something's weird about the `eslint-plugin-functional` package that makes
'react-scripts build' unable to find its rules.  Running `eslint` directly
works fine--the rule violations are ignored with the comments there, and
flagged as errors if the comments are missing, but `build` somehow runs
`eslint` in a different way that makes it fail to find the rules for that
plugin. If finds rules for other plugins, and complains in a different
way if you give it a rule for a nonexistent plugin, so it's hard to see
where the disconnect is.

There's an issue (https://github.com/jonaskello/eslint-plugin-functional/issues/105)
but it's new and has no comments yet.

This works around the problem by just using `eslint-disable-next-line`
without specifying the rule (adding a comment in each case to explain why).