Avoid flicker on selected features when updating map
Without this change, since MapBox updates are asynchronous, despite the
fact that _call_ to `map.setData` was happening before the call to
`map.removeFeatureState`, the features were being removed before the map
was updated. For adding new geounits to a district, this resulted in a noticeable
"flicker" effect whereby the selected feature state was removed before
the GeoJSON was updated with the new district, resulting in the white
map beneath being shown and then the district with the appropriate color
being filled in shortly thereafter (hence the flicker).

With this change in place, when adding or edit district geounits, the
selected feature state is only removed _after_ the GeoJSON has already
been updated, avoiding the white map being shown, reducing the flicker.
When removing a geounit from a district we continue to remove the
feature before updating the GeoJSON.