Deselect features using rectangle tool
This makes it so that features are both selected _and_ deselected
properly on mouse move (deselection was previously not working).

Originally, the feature selection was only happening after the user had
finished drawing the rectangle. When making the change to dynamically
select features as the mouse moved (as rectangle was resized),
deselection was neglected (not implemented).

With this change, as the rectangle is resize, features are selected and
deselected appropriately based on the overlap.