Add rectangle selection tool (#206)
* Add dependency

* Set selection tool to rectangle

* Switch to mapbox-gl-draw-rectangle-mode plugin

* Copy/paste code for mapboxgl-draw-rectangle-drag directly into project

* Go back to using mapboxgl-draw-rectangle-drag dependency

* Refer to built module file for mapboxgl-draw-rectangle-drag

* POC rectangle selection with no extra deps

* Rectangle selection/deselection working

* Make rectangle selection work via button

* Make rectangle selection only additive

* Set cursor to pointer only when hovering over layer

* Ensure event handlers are set before removing them

* Use icons for buttons

* Style buttons based on selected selection tool

* Remove lint

* Work around 'Definition for rule was not found'

See https://github.com/jonaskello/eslint-plugin-functional/issues/105

* Use constants instead of magic strings

* Reorganize code

Break common map functions out into separate file. Put each selection
tool in its own file and document behavior.

* Select features as the rectangle is drawn

in keeping with the prototype's behavior.

* Style map header

* Fix labels after rebase

* Deselect features using rectangle tool

This makes it so that features are both selected _and_ deselected
properly on mouse move (deselection was previously not working).

Originally, the feature selection was only happening after the user had
finished drawing the rectangle. When making the change to dynamically
select features as the mouse moved (as rectangle was resized),
deselection was neglected (not implemented).

With this change, as the rectangle is resize, features are selected and
deselected appropriately based on the overlap.

* Lint

* Adjust styles so label dropdown is within map header