Fix rectangle selection tool bug with multiple selections (#208)
* Fix bug with multiple rectangle tool selections

When making a selection with the rectangle tool, finishing that
selection, and then starting another one, upon mouse move the original
selection would be lost. This change makes it so that original selection
is accounted for an not deselected in error as was happening before.

* Rename variables

The term "selected" was confusing in this context because it was used in
the Mapbox sense of the feature state and also in the sense of being
within the bounding box of the rectangle. This change makes it only be
used in the Mapbox sense.

* Placate linter

* Make style change