Save as larger geounit when all values are the same
This is relevant when the following happens:
- Entire non-base geounit (eg. county or blockgroup) is selected and assigned to a district
- One geounit within it is reassigned to a different district
- Then it is later reassigned back to the original district

This change makes it so that instead of saving the data for that
geolevel as an array containing all the same value, it contains just the
value itself.

This keeps our district definitions maximally compact and also makes it
easier to identify unassigned geounits which will be relevant for future
feature work.