Avoid issue with side effect clearing features
We have an effect in `Map.tsx` which listens for there being no geounits
and then clears out the features. This becomes problematic when we want
to both add and remove features as a single, atomic operation. The
workaround in this case was first to make sure to add first and then
delete so as to not clear out all geounits "in the middle" of the
operation. The better workaround is to have the adding and removing be
part of a single state update.

All of that said, it might be better still to rework this so as to get
rid of the effect all together.