Select geometries at any geolevel (#287)
* Do not auto-accept selection when switching geolevel

* Disable base geolevel/non-base geolevel selection as appropriate

Hide higher geolevel selection options when a selection is made at the
base geolevel and also hide base geolevel selection option when a
selection is made at the non-base geolevel. Essentially, we want to
never select both base geolevel and non-base geolevel geometries at the
same time.

There is an implicit assumption baked into this logic that there is only
one base geolevel whose geometries are not seen all the time.

* Code cleanup 🧹

* Use min/max zoom from metadata for vector source

* Change tooltip to reflecting zooming in/out appropriately

* Show hardcoded constraints on hover

* Show constraints in tooltip

* Use bolded text instead of password style constraints for tooltip

* Refactor to move more functionality into GeoLevelButton

* Refactor to break tooltip out into its own function

* Fix geolevel selector styles

Wrapping the buttons caused issues with matching the :first-child and
:last-child selectors. The solution here was to add an explicit button
wrapper class and match on that.

* Use box shadow instead of border as per mocks

* Change button label from 'Approve' to 'Accept'

* Make zoom text a constant

since the user never needs to zoom out to activate any geolevel.

* Format CSS

* Allow for selecting/deselecting sub-geounits within selection

* Fit map to bounding box before querying for features

* Do nothing when switching back to larger geounit

* Do nothing when switching back to same geolevel

* Make sure fitting to map bounds has finished before modifying state/map

Also add geounits before removing them to avoid unwanted side-effects.

* Deselect sub-geounits when selecting containing geounit

* Lint

* Remove leftover console logs

* Fix error when calling findSelectedSubFeatures on blocks

* Aggregate demographics for all selected geounit geolevels

* Remove unused variable

* Avoid error when no geounits selected

* Calculate demographics from every geolevel correctly

using the geounit hierarchy.

* Avoid issue with side effect clearing features

We have an effect in `Map.tsx` which listens for there being no geounits
and then clears out the features. This becomes problematic when we want
to both add and remove features as a single, atomic operation. The
workaround in this case was first to make sure to add first and then
delete so as to not clear out all geounits "in the middle" of the
operation. The better workaround is to have the adding and removing be
part of a single state update.

All of that said, it might be better still to rework this so as to get
rid of the effect all together.

* Use imperative style for performance reasons

* Avoid unnecessarily looking for sub-features

* Use DemographicCounts in more places