Fix S3 permissions to allow CloudFront logging
It appears as though something changed with Terraform's ability to
detect out-of-band grants, or perhaps a CloudFront service level change
led to different behaviors than before (they recently added real-time
logs).

When logging to S3 from CloudFront is configured via the AWS console, a
specific set of grants that allow CloudFront to interact with an S3
bucket are added. These grants appear to get wiped away by Terraform
unless they are explicitly defined in the Terraform configuration.

The changes in this PR aim to persist the two grants (one is the AWS
canonical user ID, and the other is the CloudFront canonical user ID) in
the Terraform project.