Set standard position for map tooltip
Instead of repositioning the map tooltip based on the cursor proximity to the edge of the screen, this commit removes that logic and always positions the tooltip on the bottom right of the cursor. I had originally suggested making the tooltip respoition, but after using it, I found it annoying to have it unexpectedly jump away from where I was expecting to see it. I think we should remove this for the time being.