Style pass on map tooltip (#414)
* Cleanup style for tooltip

Tighten line-height to make it easier to compare bars in race chart, set fixed size so the size of the tooltip doesn't change as much, making it easier to quickly scan as it changes.

* Set standard position for map tooltip

Instead of repositioning the map tooltip based on the cursor proximity to the edge of the screen, this commit removes that logic and always positions the tooltip on the bottom right of the cursor. I had originally suggested making the tooltip respoition, but after using it, I found it annoying to have it unexpectedly jump away from where I was expecting to see it. I think we should remove this for the time being.