Allow saving with undo/redo (#498)
* Limit the undo history max length

* Refactor to let the map update itself after undo/redo

This makes it so that the reducer is no longer in the business of
updating the map directly. Instead, the reducer updates state that then
affects the map, but the map itself is responsible for updating itself
in response to state changes.

* Save action along with state in undo history

* Trigger effect when undoing/redoing

* Access present state correctly in components

* Remove unused import

* Avoid pushing to state after successful save

* Use helper function

* Use effect from present state

* Add separate action for when user saves districts definition

* Make replaceState preserve effect

* Simplify state snapshot interface

Discriminating based off of properties proved to be too error prone.
It's simpler to just have an optional effect and a consistent shape.

* Get undo mostly working

Passing in a specific districts definition for updates allows us to save
a custom definition, "locking in" the one we want for that specific
snapshot of state.

* Use replace state instead of updating manually after districts definition update

* Avoid overwriting effect when replacing state + comments

* Don't modify project data on undo/redo

* Don't keep track of prev/current project data

* Make effect a callback

so correct state may be injected depending on which way we're going
(undo vs. redo).

* Make it easier to patch state instead of updating/replacing entirely

* Remove console.log

* Add more comments

* Simplify type

* Correct the undo history max length

This doesn't actually matter too much since it'll be a big number but
this seems more correct.

* Update changelog

* Do not preserve effect in pushStateUpdate

* Add helper for truncating past undo history

* Move undo/redo functions and types into separate module