Avoid mixed type values in SHP export to prevent them from being dropped
The Shapefile format doesn't seem to like mixed data types, and so our
compactness column was being dropped when either the unassigned district
had a NULL value for compactness or when at least one district was
non-contiguous.

To accomodate this, I've introduced a generated "contiguity" column of
string type, and switched from using NULL to using 0 in cases of no data.