Update Nest dependency types and fix import
Nest was failing to build since `HandlebarsAdapter` is no longer
exported from the same module. The import was updated as per the current
docs: https://nest-modules.github.io/mailer/docs/mailer

`handlebars` was also added as a peer dependency since the docs said it
should be installed if it is used (there was also a warning about this).