Update Nest.js deps (#637)
* Upgrade Nest dependencies

via `nest update -f -t latest` (see
https://github.com/nestjs/nest-cli/issues/321#issuecomment-506646369)

TypeORM was also upgraded separately via `yarn upgrade --latest typeorm`
since it seems it was installed as a sibling dependency (see
https://github.com/PublicMapping/districtbuilder/commit/10fcff97ee1d66cf6c6b28d1782445fe24f2f0f7#diff-d57c900a61f2746ea3823bed028e1d81004e1e077149e4b4935269ffb39383e8)
which is also how the docs indicate it should be done (see
https://docs.nestjs.com/techniques/database).

* Update Nest dependency types and fix import

Nest was failing to build since `HandlebarsAdapter` is no longer
exported from the same module. The import was updated as per the current
docs: https://nest-modules.github.io/mailer/docs/mailer

`handlebars` was also added as a peer dependency since the docs said it
should be installed if it is used (there was also a warning about this).

* Downgrade from @nextjs/crud alpha version to stable

* Downgrade to avoid regression in minor version of @nestjsx/crud

When moving to 4.6.1+ an error is thrown at runtime when we try to list
projects:

server_1    | QueryFailedError: column Project.undefined does not exist

This is related to the filtering of the project controller. In looking
at the WHERE condition in the query to fetch the projects we can see
where the query is being built incorrectly (a column
`"Project"."undefined"` is referenced):

server_1    | query failed: SELECT "Project"."id" AS "Project_id",
"Project"."name" AS "Project_name", "Project"."number_of_districts" AS

...

"regionConfig"."id"="Project"."region_config_id"  INNER JOIN "user"
"project_user" ON "project_user"."id"="Project"."user_id" WHERE
("Project"."undefined" = $1) ORDER BY "Project"."updated_dt" DESC --
PARAMETERS: ["07f9966f-facf-42ce-a4ee-48320007f1cc"]

It seems like this should work fine since `user_id` is indeed a column
on the project.

An issue was created for this here:
https://github.com/nestjsx/crud/issues/684

* Finish upgrading all Nest.js deps

This makes it so that all deps which can be upgraded are upgraded.

`class-transformer` had to be held back because upgrading it caused
`@nestjsx/crud-typeorm` to throw an error since it seems to depend on an
older version.

* Pin to exact versions

* Update changelog