Limit drawing to within starting county
- Add changelog

* Move draw options next to drawing tools

The "limit drawing to ..." type options are related to the drawing tools, so this commit moves the
button next to the drawing tools.

* Change positioning for menu

By adding a container element with position: relative, we are able to position the menu relative to
the container of the button, rather than using fixed position to position it depending on the global
page positioning. The previous styles was brittle and would break if the options button moved or if
other elements on the page changed size.

* Fix style and user experience for "limit draw to county"

Fix lint

Add comment explaining function