Use setPaintProperty to updated district borders on select/hover
This makes it _much_ snappier to update the district borders since we
don't have to update all the feature properties and re-set the map
GeoJSON to change layer styles.

The downside to this is it we're now setting map styles both
declaratively via feature properties and via updating map styles
imperatively "out of band" in a separate place in the code (separate
effect). This can result in subtle interplay between the two, such as a
discrepancy between the feature properties and the styles shown on the
map. I think in this case the added complexity/maintainability burden is
warranted since this will be a very often used feature by all users.