Change positioning for menu
By adding a container element with position: relative, we are able to position the menu relative to
the container of the button, rather than using fixed position to position it depending on the global
page positioning. The previous styles was brittle and would break if the options button moved or if
other elements on the page changed size.