Make password validation visible in modal
The password validation panel was being cut off by the modal's `overflow: hidden` property. I
removed `overflow: hidden` from the modal. I also reformatted the validation panel to take up less
space, which should help us avoid situations where it would expand to the edge of the browser
window.

I experimented with rendering this panel above and below the field, but in both cases it blocked and
prevented interaction with other important elements on the page. I think this solution will work
well in a majority of cases.