Highlight selected/hovered districts and zoom to district (#688)
* Show outline for selected district

* Show border around district when hovering row in sidebar

* Add district options flyout with zoom to district

* Update changelog

* Double the thickness of selected/hovered district outline

* Use middle dot character for hamburger icon

Co-authored-by: Michael Maurizi <mmaurizi@azavea.com>

* Add missing table header for new column

* Style fixes

Make district flyout button more in keeping with the styles for the lock
button (tooltip, same circular shape and background color). Also move
flyout menu slightly to the left so there's no "gap" in which the menu
disappears.

* Use onMouseEnter and onMouseLeave for setting hovered district

This results in less events being fired since child events don't bubble
up with these handlers (and we don't want them to!). Previously, mousing
over any child (eg. the lock district button) would trigger an
`onMouseLeave` event.

* Use setPaintProperty to updated district borders on select/hover

This makes it _much_ snappier to update the district borders since we
don't have to update all the feature properties and re-set the map
GeoJSON to change layer styles.

The downside to this is it we're now setting map styles both
declaratively via feature properties and via updating map styles
imperatively "out of band" in a separate place in the code (separate
effect). This can result in subtle interplay between the two, such as a
discrepancy between the feature properties and the styles shown on the
map. I think in this case the added complexity/maintainability burden is
warranted since this will be a very often used feature by all users.

* Style cleanup for highlight selected district (#703)

* Fix district line render order on map

The district lines should be rendered beneath label layers.

* Style cleanup for "zoom to district"

Adds a new menu component that works properly with scrolling content. The previous menu did not
render in the proper position when the sidebar was scrolled. Small cleanup on the button, added an
ellipsis icon.

Not yet working:

- The "Zoom to district" button doesn't currently function
- When a user clicks on the overflow button on a non-selected district, the district gets selected;
  ideally, only the menu would open

* Fix "zoom to district" functionality

Co-authored-by: Peter Caisse <pcaisse@gmail.com>

Co-authored-by: Peter Caisse <pcaisse@gmail.com>

* Remove unused imports/type

Co-authored-by: Michael Maurizi <mmaurizi@azavea.com>
Co-authored-by: Jeff Frankl <jfrankl@azavea.com>