Log non-HTTP errors to Rollbar
Our logic was a little flawed in that we required errors to be
`HttpException`s to be logged to Rollbar. However, we also want to know
about other types of unhandled exceptions.

This is slightly tricky in terms of types since in JavaScript, bizarrely, you can
actually throw _any_ type (does not need to be an `Error` or extend
`Error`). However, that seems to be enough of an edge case that we don't
have to worry about it.

In the specific case of the leaving-an-organization bug, that generated
a `QueryFailedError` which is why it was not sent to Rollbar.