Style cleanup (#904)
* Style cleanup on community maps page

Position 'filter by state' in top right of screen.

* Style fixes on 'my maps' page

* Style cleanup on 'new map' view

Set consistent padding between 'Organization' and the other sections. Add box shadow to match other components in the application. Improve consistency across cards.

* Cleanup style for 'Expanded Sidebar'

Fixes #865.

* Style cleanup for 'Evaluate Mode'

Make the 'Send to PlanScore' button fixed on the screen, so it's always visible. After clicking that button, the opacity is higher, making it easier to read in its disabled state. Make the Demographics column of 'Majority-Minority' left aligned, since it's not a number.

* Style cleanup for 'find' button

* Remove unused prop

* Remove unused prop

* Increase size of maps on My maps view.

* Fix inconsistent copy

* Clean up on New Map view