Create auto scaling group and stand up new app service
I made the following changes to support allocating over 30 GB of memory to
the app server for the updated Census:

- Create an Auto Scaling group and launch template based on the Terraform
  0.12 branch of our ECS cluster module.
- Make changes to work with the existing ECS cluster and security group
  resources.
- Stand up a second ECS service for the application server that uses the EC2
  launch type.
- Make use of the awsvpc network mode to preserve the existing target group,
  allowing a zero-downtime deployment.

In previous ECS clusters where we've managed container instances, we've used
bridge networking. awsvpc is the same network mode that we use with Fargate,
which makes it less complicated to bridge two launch types, but also means
we must have a one-to-one mapping of application servers to container
instances.