Fix importing to a region with identical geounits at different levels
For PA LRC, we created a region with block>precinct>municipality>county
as our set of geounits. Because Philadelphia is a city-county, that means
the county-level will have an array with one element (the municipality-level).

This threw off the code for simplifying newly generated districts, which
assumed that all array items in the districts definition have more than 1
element, leading it to incorrectly "simplify" the districts definition by
merging the county & municipality levels for Philadelphia