Calculate population to be added correctly
The tooltip in the deviation column was incorrectly calculating the
population that needed to be added to a district that was below the
population deviation threshold. Instead of subtracting the current
deviation from the maximum allowable deviation, it was being added,
resulting in instructions to add very large population numbers in order
to hit the target being displayed in the hover tooltip.