Add logic check to prevent yScale from generating negative heights
The yScale function was trying to generate bar heights before the
page was fully loaded and parent.height had been assigned a value,
creating "negative values are not valid" errors in the console.
This commit adds logic to check whether parent.height exists, returning
the upper end of the yScale range (10) if it does not to prevent scaling
in these cases.