Switch PlanScore integration from simple to multi-step API
The PlanScore backend has a file size limitation for direct upload, but
allows any size file if you instead upload your file directly to S3 using
a signed URL.

I initially attempted to do this in a "stateless" fashion, hitting our
backend for API requests to PlanScore but posting to S3 directly from the
browser, which would have eliminated the need to do polling on the client.
Unfortunately I couldn't get this to work, due to both CORS issues and
complications with using the callback approach for the S3 upload.