Move worker timeout check into the workerQueues task
With the timeout checking done outside of the queue, timed out tasks
have their worker thread recreated, but no new tasks can be queued because
the timed out task will never complete or be rejected.

By moving the timeout into the body of the workerQueue task, when we return
an error that task will be removed from the queue, allowing future tasks to
be run instead of blocking forever.