Add Github Release workflow with plan only
This was done in the same way to match other Azavea projects (not public)
and to match the way Jenkins works for the project thus far as well.

Entering a sha with a valid image in ECR is therefore required.

This adds a release workflow that can only plan a production change.

Related: #1237