Upgrade to VPC that drops bastion and remove other related bits
The bastion has been replaced by using AWS SSM and the use of PGadmin was
confirmed to work in an earlier commit.

As part of upgrading to the new VPC we had to upgrade Azavea providers and cert
versions so that AWS providers versions didn't conflict.

Note from deploying: had to manually remove the bastion security group from the inbound one