Enable SSM to ec2 container hosts
This allows us to move away from using the bastion and
to connect to a host by just pressing Connect on the instance.

Also installed psql on the host so we can connect to database too.

It should be noted that these hosts can get recycled as part of deployments -
so that's the only process change that needs to happen.