Upgrade to React 17
The current version of theme-ui isn't compatible with the new React
JSX transform, so we needed to add DISABLE_NEW_JSX_TRANSFORM=true to the
client container environment to disable it.

Upgrading to later versions of theme-ui resulted in numerous Typescript
errors. We're going to have to upgrade theme-ui to the latest version in
order to upgrade to React 18 anyway, and theme-ui 0.15.0 contains
breaking changes that I think will incidentally fix the Typescript
errors I was seeing (we're on 0.13.1 and the errors starting showing up
with version 0.14.0, so it didn't seem like the best use of time to look
into fixing them at this point.

Anyway, we should remove DISABLE_NEW_JSX_TRANSFORM=true from
docker-compose.yml when we upgrade to React 18.