Refuse to load archived RegionConfig in TopologyService.get
This shouldn't get called for archived regions, but somehow it is, and it causes
big problems when the region in question is broken.
This will make it bail out rather than try to load an archived region, which
should limit the scope of the problem to whatever old entity is trying to
load it, rather than breaking the whole site.